backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg[binary]
python-dotenv
django-redis      
celery            
redis             
